Initial code for DVB subtitles that simply dumps the subtitles stream.
This is intended as an initial testing code for opening a subtitles PES
from the hardware demuxer and reading out the specific DVB subtitle
elementary stream.
For testing this initially just dumps a hardcoded subtitle PID to file.

With this I was able to get a binary dump of the data that will flow around
to process it further. Seems to need some parsing out some metadata (PES packet
information and subtitle segment headers I think?) before passing on to
dvbsub_decode ported from libavcodec.