Initial parsing code.
Initial parsing code that is capable of walking the full content of a
subtitle PES data stream dump and output information on a lot of its
contents.
The PES header is parsed for the private data length and the
segment type is detected. For page composition and object data some
additional details is printed out for testing. End of display set
segments are reported as well.
Added test code for feeding it a dump and calling the functions for
testing.
Hooked up to autotools.