Some code tweaks while debugging the unknown pixel block reasons.
* Compare region with NULL in the return check, instead of 0, as region variable is a pointer.
* Enable the packet contents debug output for now, it's quite nice hexdump.
* Add some FIXME's for memory safety checking

Note that the unknown or unsupported pixel block warnings happened indeed simply
because read_nbit_string functions aren't implemented yet, but not because of
their return value, but because an address to the buffer is given to the function,
which in turn walks it on its own, so that when returned, the buf pointer has been
increased by the amount of code parsed by the per-depth function, and so the
next switch statement will work on the start of the next data_type. Without the
implementation it will try the immediate next byte, which is not data_type, but
part of the nbit/pixel_code_string. So after implementing those readers, it should
solve it nicely.