Implementation of _dvb_sub_read_4bit_string, ported from libavcodec
This is the case that seems to be used in the Finnish channel subtitle
dump I use for testing, so implemented that first, and 2bit and 8bit
cases will follow later (probably after the rest works).
It uses GstBitReader for easier porting and consequently makes the
library depend on libgstbase-0.10.so (part of gstreamer core package).