Implementation of end of display set parsing, ported from libavcodec.
This also adds saving to *.png files when enabled with a macro.
It shows that things don't seem to quite work yet in read_4bit_string

The representation of the final data is currently ported straight from
libavcodec representation. We probably will want to change this to
something more suitable for Qt/gstreamer straight away, as I can see
already many ways to improve the data structures, both for sanity and
suitability for conversion to gst/QtImage