Fix a bug in the code ported from ffmpeg relating to string reading loop.
The code in libavcodec checks for bits remaining to be less than buf_size,
but in my test sample they are always exactly equal, and the loop is never entered.
This can't be right, so fixed to a less than or equal check.

Now at least some string data is tried to be parsed, but still without getting good results.
It seems it is overwriting from position 0x0 and 0x1 (top or bottom half of the interlace being
handled), and that's what needs to be debugged next. y_pos doesn't seem to get incremented
properly. Related to that, the debug purpose resulting image dumping is enabled until all
current rendering bugs are squashed.