Convert object and pixel parsing prints to dvb_log.
Also changes log type strings to shorter capitalized versions, as these
are used for entering from environment variable, and shouldn't be so long.
Later we can convert it to a table, where a longer string for debug log
prefixing could be added.

Also adds a few more debug messages that I've had locally for a while for
the pixel block parsing problem debugging.