Fix a crash with debugging enabled, other debugging changes.
A NULL pointer was derefenced in debugging print, because map_table
was probably unset as we were in that if else block, so print the
correct variable.
Remove linebreaks from the end of debug messages that are using dvb_log
now, as that adds its own linebreak already, and having an empty line
is not helpful for debugging.
Comment out some memory content dumps that spam too much and aren't
interesting right now.