Refactor read_4bit_string completely.
Add an alternative implementation of read_4bit_string that's a
refactoring of the previous (which is kept #if 0'ed out for the immediate
time being).

The previous code (straight ported from ffmpeg) was doing certain things in
almost all branches with lots of code duplication, complicating things too
much for debugging. These repeating things were:
* handling overflow of the line (pixels_read must be smaller or equal than dbuf_len)
* mapping of parsed index to map_table index if necessary
* handling of non_modifying bytes. If non_mod bit was set in object segment (passed
  to us as an argument) and bit is "1" (the non_mod bit if non_mod is used in the segment),
  then it was skipped in many different places
* actual setting of the destbuf bytes to the determined bits (CLUT index really)

This is done by simply gathering the runlength and clut_index (previously kept as color
index in often used "bits" variable) into variables in the runlength algorithm branching,
and just doing all the above common things at the end of the loop, after the branches.
All branch trees are either setting both runlength and clut_index, keeping clut_index
as zero or keeping run_length as zero (in which case all the other parts are not done
at the end of the loop, this is used for stop_parsing case too)

As part of the refactoring, the following improvements (in bad data handling mostly)
are achieved with some different approches:

* If the input data is bad, in the sense that more pixels are present in the
  4-bit/pixel_code_string than the actual region width is, the excess bits in
  source data is consumed as well, but gets ignored
  (run_length = MIN (run_length, dbuf_len) should take care of that).
  Before the refactoring, the loop exited immediately when all pixels were read,
  and if the stream for some reason had more pixels defined than the region
  allows (could be a different bug showing its head, but we need to be safe from
  such ill-formed input data anyway) and the end_of_string_signal needed extra
  consumption after the loop. With the refactoring both of these aspects just work
  naturally.
* The setting of clut indeces is done with a memset, which could be quicker in
  case of larger runlengths (depending on glibc arch specific memset routine
  optimizations), than the previously used while loops that set it byte by byte.
* Better code maintainability, debuggability and so on of course.