Fix the boundary checks for parse_pixel_data_block
The condition on which a line data overflow was detected was
(x_pos > region->width || y_pos > region->height),
but x_pos and y_pos are zero-based, while width and height are not
quite zero-based - they are width/height in pixels, so if
y_pos is 42 and height is 42, then we are already overflowing because
really the highest (visually lowest) allowed y_pos is 41, as it starts
from zero.

So modify the check to
(x_pos >= region->width || y_pos >= region->height)
to not feed data into read_nbit_string with a destination buffer that
is out of its bounds, or it will scribble over memory we don't own,
eventually leading to heap corruption and crashes -- commonly seen
crash was related to dvb_log freeing some encoding conversion data
via g_logv mechanisms that was corrupted by the memset in read_4bit_string
over the bounds due to this bug.

Also add some more verbose debug information.

This bug seems to be present in FFMPEG avcodec dvbsub decoder as well,
from which it was ported to here.