Relocate and relax the returning when destination buf is out of bounds.
Since we now correctly check for y_pos >= region->height, we get the
warning and early return even if we have valid data that has all
pixel data correctly with an end_of_object_line in the end, because of
that end_of_object_line handling being in the same switch.

So move the actual warning and returning to the 2/4/8 pixel string region
switch cases, but keep the check before the switch to reduce code size,
using a dest_buf_filled boolean to track it. We'll probably use the boolean
later on anyway to pass dbuf_len=0 to read_nbit_string instead of warning
and returning.

This also removes the x_pos >= region->width case, because we should be
able to assume that dbuf_len will be nicely passed as zero to read_nbit_string
anyway, and we don't want the dest_buf_filled boolean to be turned TRUE for
these cases, as it's just typically awaiting a end_of_object_line case, to
proceed to the next line perfectly by spec.

Also add some debugging on the invalid object data cases in the form on
dumping the memory contents of the remaining data for now.