Add a FIXME about a bug issue that leads to walking over memory we don't own.
For some reason it is currently possible that we don't get a full PES packet, so
I noticed a case where the PES_packet_length was claimed to be over 5000bytes,
but the data we were parsing was just a bit over 200 bytes. This lead to a lot
of walking over bad data with a huuuuge number of warnings, taking a lot of time
to print out.
We will want to at the very least guard the data length passed in with the minimum
of available data and the claimed remaining PES packet length, or perhaps buffer
data up if not full PES length and gathered it up from next feed calls.
Best if we could ensure DVB demuxer isn't giving us half packets as I believe
it should behave like.