Don't use GSList for DvbSubPrivate::object_list DVBSubObject list
Simply use the (previous unused but accidentally remained from ffmpeg ported code)
next pointer in DVBSubObject structure, the GSList abstraction
just complicates list walking and freeing for no real benefit here.

With this get_object and the creation is equal to how it is in ffmpeg, except
for using glib memory allocation instead of avutil of course.