Initial delete_state implementation ported from ffmpeg
Uses the g_slice_free_chain convenience function in case of the clut
list, instead of walking the chain itself. This goes directly to
GSlice level as well and might be more efficient, as it doesn't need
to get to the GChunk slabs or magazine layers for each g_slice_free
in the loop, but needs to get those just once.