Implement subtitle testing fully for gstreamer dump source case.
We were passing the data to dvb_sub_feed_with_pts, but that assumes
that we are passing only one PES packet to it for obvious reasons, so
loop it ourselves until an error or when all data is consumed.

For this to work properly, dvb_sub_feed_with_pts needs to return the
correct number of bytes it consumed, which will be committed in a
subsequent commit.