Implement _dvb_sub_read_8bit_string fully, but blindly.
I have no test subtitles that would use this 8bit string coding method,
so this is just written by the spec.
Not using code ported from ffmpeg for this, because it is unsafe in face
of bad data.

Uses the same algorithm method as refactored read_2bit_string and
read_4bit_string.