Wrap open/ioctl calls between TEMP_FAILURE_RETRY and open demuxer RDONLY.
TEMP_FAILURE_RETRY is a unistd.h defined macro, which retries executing
the expression until it isn't returning -1 with errno as EINTR.
This is necessary if some signal handler in the same thread wasn't
installed with the SA_RESTART for some reason. So be resilient against
that, just because we can.

We are only reading from the dvb demuxer, so open RDONLY instead of RDWR.