API change: Include the PTS in the new_data callback for syncronization with video purposes
The library user needs to know when to actually display the subtitle set,
as in, be in sync with the displayed video stream as the video is typically
buffered up a bit, while the subtitle stream is not, so without this the
library user can't do anything about subtitles getting displayed too early.
Also, the specification says to display the subtitles at PTS, and gives no
guarantees that the subtitles aren't described many seconds before the actual
point in the video stream in the container, but only that PTS is monotonically
increasing in the subtitle stream.